Rag-chatbot-assistent/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
langchain>=0.3.7
langchain-google-genai>=2.0.4
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
langchain-core>=0.3.15
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
tiktoken==0.8.0
pydantic==2.10.2
numpy>=1.26.0
